MP-152 :truck: Rename: 서비스 인터페이스들 이동
- 도메인에 CRUD가 제외됨에 따라 domains > commons > domain > supers에 있던 항목들을 domains > commons > app > service > supers로 이전한다.
