+++ src/main/java/kr/modusplant/domains/commons/app/service/supers/UuidCrudService.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.commons.domain.supers;
+package kr.modusplant.domains.commons.app.service.supers;
 
 import java.util.Optional;
 import java.util.UUID;
