+++ src/main/java/kr/modusplant/modules/jwt/domain/service/supers/RefreshTokenCrudService.java
@@ -1,6 +1,6 @@
 package kr.modusplant.modules.jwt.domain.service.supers;
 
-import kr.modusplant.domains.commons.domain.supers.UuidCrudService;
+import kr.modusplant.domains.commons.app.service.supers.UuidCrudService;
 import kr.modusplant.modules.jwt.domain.model.RefreshToken;
 
 import java.util.Optional;
